requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn==1.5.1
joblib
pydantic
openai>=1.0.0
python-dotenv